requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.7.1
python-dotenv==1.0.1
pdfplumber==0.11.1
requests==2.32.3
google-genai==1.5.0
langchain==0.2.1
langchain-core==0.2.1
langchain-google-genai==1.0.3
